+++ Modules.php
@@ -77,9 +77,12 @@ if ( ! $allowed )
 
 				// Eg: "Student_Billing/Statements.php&_ROSARIO_PDF".
 				$_ROSARIO['ProgramLoaded'] = $program;
-
-				break 2;
 			}
+		}
+
+		if ( $allowed )
+		{
+			break;
 		}
 	}
 }